backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pandas==2.2.3
numpy==2.2.1
openpyxl==3.1.5
anthropic==0.42.0
python-multipart==0.0.20
matplotlib==3.10.0
python-dotenv==1.0.1
aiofiles==24.1.0
